use "git config --edit" instead of manually launched editor
This fix is intended for Windows. Because GIT_EDITOR variable and GIT_DIR
are both adapted for unix environment, it is tricky to recover from them
command line to run editor from native Windows application.

Fix it by using git to run the editor itself.

Downside is that if there is dedicated `sequence.editor` value different
from commit message editor it would not be user. I think it is
not important.